move encrypter init to initializedconfig method
